refactor: add willo and sokoban modules and move relevant items to them (#61)
* refactor: move willo-related systems out of sugar module

* refactor: move core player components and events to willo module

* chore: actually add willo module

* refactor: move general grid-based systems to sokoban module

* refactor: move move_player_by_table system to willo module

* refactor: move player_state_input system to willo module

* refactor: move PlayerBundle and more player-specific components to willo module

* refactor: move WallBundle to sokoban module

* refactor: rename player items to willo items

* refactor: move willo app logic to new WilloPlugin

* refactor: move sokoban app logic to new SokobanPlugin

* refactor: encapsulate many items in willo module

* refactor: encapsulate many items in sokoban module

* docs: add some basic docs to the remaining public willo items

* docs: add basic documentation to the remaining public sokoban items
