+++ src/gameplay/mod.rs
@@ -36,13 +36,8 @@ impl From<Direction> for IVec2 {
 }
 
 #[derive(Copy, Clone, Eq, PartialEq, Debug, Hash)]
-pub struct PlayerMovementEvent {
-    direction: Direction,
-}
-
-#[derive(Copy, Clone, Eq, PartialEq, Debug, Hash)]
 pub struct DeathEvent {
-    pub player_entity: Entity,
+    pub willo_entity: Entity,
 }
 
 #[derive(Clone, Eq, PartialEq, Debug)]
